requirements.txt
django~=4.2
Pillow
python-dotenv~=1.0.0
psycopg2
gunicorn
djangorestframework~=3.14.0
djangorestframework-jwt~=1.11.0
django-cors-headers